Add Entity models for account and transaction tables.
Update properties.
Add account repository and inject into service.

+++ money-market-shoprite/src/main/kotlin/com/example/moneymarketshoprite/models/TransferCommand.kt
@@ -1,5 +1,5 @@
 package com.example.moneymarketshoprite.models
 
-class TransferCommand(val depositAmount: Double, val currency: String, destinationAccountNumber: Int)
+class TransferCommand(val depositAmount: Double, val currency: String, destinationAccountNumber: Long)
 
 //Currency to be Value Object
